+++ test/pages/check-beacon-summary.test.tsx
@@ -1,10 +1,18 @@
 import React from "react";
 import { render } from "@testing-library/react";
-import CheckBeaconSummaryPage from "../../src/pages/check-beacon-summary";
+import CheckBeaconSummaryPage from "../../src/pages/register-a-beacon/check-beacon-summary";
 
 describe("Check Beacon Summary Page", () => {
   it("renders correctly", () => {
-    const { asFragment } = render(<CheckBeaconSummaryPage />, {});
+    const testBeacon = {
+      beaconManufacturer: "Raleigh",
+      beaconModel: "Chopper",
+      beaconHexId: "00FF00",
+    };
+    const { asFragment } = render(
+      <CheckBeaconSummaryPage {...testBeacon} />,
+      {}
+    );
     expect(asFragment()).toMatchSnapshot();
   });
 });
